Conflicts:
	pipeline/master/bbs.py
	pipeline/master/mwimager.py

git-svn-id: http://usg.lofar.org/svn/code/trunk@3500 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/pipeline/pipeline/master/bbs.py
@@ -37,7 +37,7 @@ class bbs(LOFARrecipe):
             '-f', '--force',
             dest="force",
             help="Automatically clean control database, sky model parmdb, and instrument model parmdb",
-            default=True
+            action="store_true"
         )
         self.optionparser.add_option(
             '--db-host',
@@ -107,15 +107,18 @@ class bbs(LOFARrecipe):
 
         try:
             self.logger.info("Running BBS")
-            self.logger.info("Executing: %s" % " ".join(bbs_cmd))
-            with closing(open(log_location, 'w')) as log:
-                result = check_call(
-                    bbs_cmd,
-                    env=env,
-                    stdout=log,
-                    stderr=log,
-                    cwd=working_dir
-                    )
+            self.logger.debug("Executing: %s" % " ".join(bbs_cmd))
+            if not self.inputs['dry_run']:
+                with closing(open(log_location, 'w')) as log:
+                    result = check_call(
+                        bbs_cmd,
+                        env=env,
+                        stdout=log,
+                        stderr=log
+                        )
+            else:
+                self.logger.info("Dry run: execution skipped")
+                result = 0
             return result
         except CalledProcessError:
             self.logger.exception("Call to BBS failed")
